--- app/frontend/src/clients/IRpc.ts
@@ -1,7 +1,0 @@
-export interface IRpc {
-  request(
-    service: string,
-    method: string,
-    data: Uint8Array
-  ): Promise<Uint8Array>;
-}
