--- app/frontend/src/clients/IServiceConstructor.ts
@@ -1,7 +1,0 @@
-import { IRpc } from "./IRpc";
-
-interface IServiceConstructor<T> {
-  new (rpcImpl: IRpc): T;
-}
-
-export { IServiceConstructor };
